fix: Make static catalog JVM-only to support Scala.js compilation
- Move StaticCatalogProvider to platform-specific directories (.jvm/.js/.native)
- Create platform-specific CatalogLoader to handle file I/O differences
- JVM: Full implementation with java.nio.file support
- JS/Native: Stub implementations that return InMemoryCatalog
- Move file I/O dependent tests to JVM-only directory
- Update documentation to clarify platform support
- Remove unused imports from shared tests

This ensures Scala.js and Scala Native compilation succeeds while maintaining
full static catalog functionality on JVM platform.
