+++ CLAUDE.md
@@ -45,6 +45,14 @@ sbt scalafmtCheck
 ./sbt "runner/test"
 ./sbt "langJVM/test"
 
+# Test compiler for Scala.js
+./sbt "langJS/test"
+
+# Compile src and test for individual platforms
+./sbt "projectJVM/Test/compile"
+./sbt "projectJS/Test/compile"
+./sbt "projectNative/Test/compile"
+
 # Run specific test class
 ./sbt "runner/testOnly *BasicSpec"
 
@@ -147,6 +155,7 @@ test _.output should be """
 - `.wv` files: Wvlet query language source files
 - Scala sources: Follow standard Maven/SBT directory structure
 - Cross-platform code: Use `%%%` for multi-platform dependencies
+- **Platform specific code needs to be placed in .jvm/src/main/scala, .js/src/main/scala, .native/src/main/scala folders**
 
 ### Code Style
 - **Scala 3**: Latest Scala version (check `SCALA_VERSION` file). No Scala 2 support needed.
@@ -164,8 +173,9 @@ test _.output should be """
 
 ### Multi-Platform Considerations
 - Use `%%%` for cross-platform library dependencies (JVM/JS/Native)
-- Avoid platform-specific APIs in shared code (`wvlet-api`, `wvlet-lang`)
+- **To support Scala.js, which has no file I/O support, avoid using java File I/O modules**
 - Native builds require specific C library dependencies
+- **In Scala.js code, avoid using Java-specific libraries**
 
 ### Performance
 - Parser uses efficient TokenBuffer for lookahead
@@ -208,6 +218,7 @@ The project follows semantic versioning and uses SBT plugins for cross-platform 
 ### Code Reviews
 
 - Gemini will review pull requests for code quality, adherence to guidelines, and test coverage. Reflect on feedback and make necessary changes.
+- To ask Gemini review the code change again, comment `/gemini review` to the pull request 
 
 ### Development Workflow
 
@@ -220,3 +231,12 @@ For error reporting, use WvletLangException and StatusCode enum. If necessary er
 ## Deployment and Documentation
 
 - For new features, update the documentation at website/docs folder
+
+## Development Checklist
+- Before commiting changes, confirm compilation passes for src/main, src/test, and Scala.js
+
+## Commit Guidance
+
+- **Include CLAUDE.md changes as needed to the commit**
+  - If modifying project structure, development processes, or adding new guidelines, update this file
+  - Ensure the guidance remains clear, concise, and helpful for developers
