+++ wvlet-lang/.jvm/src/main/scala/wvlet/lang/catalog/StaticCatalogProvider.scala
