+++ wvlet-lang/.jvm/src/test/scala/wvlet/lang/catalog/StaticCatalogProviderTest.scala
