+++ wvlet-lang/.jvm/src/test/scala/wvlet/lang/compiler/StaticCatalogCompilerTest.scala
